Separate LiveStream logic from Coronavirus logic
- Move livestream actions out of Coronavirus controller into
a new controller.
- This will make it easier to check if a user has the necessary
signon permissions to update the livestream.
