+++ app/views/live_stream/index.html.erb
@@ -21,7 +21,8 @@
           <p>Click on the video, and copy its URL.</p>
         </div>
       </details>
-      <%= form_with(url: "update_live_stream", method: "put", local: true) do %>
+
+      <%= form_with(url: live_stream_path(@live_stream), method: "put", local: true) do %>
         <%= label_tag(:url, "YouTube URL", class: "govuk-label" ) %>
         <%= text_field_tag(:url, nil,
           class: "govuk-input govuk-!-margin-bottom-6",
@@ -46,7 +47,7 @@
         padding: true
       } %>
       <div>
-        <%= link_to "Publish", coronavirus_publish_live_stream_path, method: :post, class: "govuk-button" %>
+        <%= link_to "Publish", live_stream_publish_path(@live_stream), method: :post, class: "govuk-button" %>
       </div>
       <%= render "govuk_publishing_components/components/heading", {
         text: "4. Check the live landing page",
