+++ spec/features/coronavirus_page_spec.rb
@@ -101,30 +101,4 @@ RSpec.feature "Publish updates to Coronavirus pages" do
       and_i_see_a_message_telling_me_that_the_page_does_not_exist
     end
   end
-
-  context "Live stream updates" do
-    before do
-      given_i_am_a_coronavirus_editor
-      stub_coronavirus_publishing_api
-      stub_youtube
-    end
-
-    scenario "Publish a valid livestream url" do
-      when_i_visit_the_publish_coronavirus_page
-      and_i_select_live_stream
-      i_am_able_to_update_draft_content_with_valid_url
-      and_i_see_live_stream_is_updated_message
-      and_i_can_check_the_preview
-      and_i_can_publish_the_url
-      and_i_see_a_link_to_the_landing_page
-    end
-
-    scenario "Adding an invalid livestream url" do
-      when_i_visit_the_publish_coronavirus_page
-      and_i_select_live_stream
-      i_am_able_to_submit_an_invalid_url
-      and_i_see_the_error_message
-      and_nothing_is_sent_publishing_api
-    end
-  end
 end
